refactoring project structure: power usage methods
