--- classes/PowerUsage/index.js
@@ -1,7 +1,0 @@
-const DeviceRaw = require('./DeviceRaw');
-const CurrentMonth = require('./CurrentMonth');
-
-module.exports = {
-  DeviceRaw,
-  CurrentMonth,
-};
